traefishing/requirements.txt
# Core dependencies
fastapi==0.95.2
uvicorn==0.22.0
jinja2==3.1.5
python-multipart==0.0.20

# Image processing
opencv-python-headless==4.8.0.74
Pillow==9.5.0
numpy==1.23.5
qrcode==7.4.2

# Mapping
folium==0.13.0
branca==0.3.1

# Environment and utilities
python-dotenv==1.0.0
requests==2.31.0

# Authentication
supabase==1.0.3
pyjwt==2.6.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# LangChain dependencies
langchain==0.0.335
langchain-core==0.1.22
langchain-community==0.0.16
langchain-huggingface==0.0.2
faiss-cpu==1.7.4

# Notifications en temps réel
sse-starlette==1.6.5